Extract common types and functions to utils from OrdersContext
