+++ src/types/transaction.ts
@@ -1,0 +1,2 @@
+export type transactionType = [price: number, size: number];
+export type transactionWithPriceType = [price: number, size: number, total: number];
